+++ plugin/android/src/main/java/expo/modules/adapters/braze/BrazeApplicationLifecycleListener.kt
@@ -1,13 +1,19 @@
 package expo.modules.adapters.braze
 
 import android.app.Application
-import android.content.Context
 import expo.modules.core.interfaces.ApplicationLifecycleListener
+import com.braze.Braze
 import com.braze.BrazeActivityLifecycleCallbackListener
+import com.braze.enums.BrazeSdkMetadata
+import java.util.EnumSet
 
 class BrazeApplicationLifecycleListener() : ApplicationLifecycleListener {
     override fun onCreate(application: Application) {
         super.onCreate(application)
+        Braze.addSdkMetadata(
+            application,
+            EnumSet.of(BrazeSdkMetadata.GRADLE, BrazeSdkMetadata.EXPO),
+        )
         application.registerActivityLifecycleCallbacks(BrazeActivityLifecycleCallbackListener())
     }
 }